rendering bug fixes. + rendering fixes around edge case when an image is at the edge of the display

+++ src/ssd1289SysCall.h
@@ -3,8 +3,8 @@
 #include <stdint.h>
 
 typedef struct {
-  uint16_t x;
-  uint16_t y;
+  int x;
+  int y;
   uint16_t w;
   uint16_t h;
 } Rectangle;
